requirements.txt
flask
requests
pyjwt
jproperties
configparser
boto3
mlflow
Flask-OIDC
pyjwt